86cym3xbe: Moving Files to proper directory
